requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.12.7
python-dotenv==1.0.0
python-telegram-bot==13.13
pytz==2023.2
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.2
tzdata==2023.2
tzlocal==4.3
